Assistant checkpoint: Fix compilation errors and code quality issues
Assistant generated file changes:
- src/main.cpp: Add missing Wire library include, Fix SD card API usage
- src/tasks/ai_task.cpp: Fix type casting issues in min() functions, Fix type casting in learning_progress, Fix type casting in update_learning_metrics, Fix remaining min() type issues
- include/ai_states.h: Create missing AI states header file
- include/system_monitor.h: Create system monitor header file
- src/tasks/system_task.cpp: Add missing includes and fix SD card API, Fix SD card type checking, Add missing system monitor function implementations
- src/ai_states.cpp: Add ai_state_to_string function implementation
- src/tasks/ui_task.cpp: Create UI task implementation
- src/tasks/scan_task.cpp: Create complete scan task implementation
- include/config.h: Add missing constants and improve configuration

---

User prompt:

Conduct a thorough inspection of the C++ project for errors, bottlenecks, and code smells.

Instructions: Begin by reviewing the project structure and organization to ensure a logical flow and adherence to best practices. Identify and document any compilation errors or warnings that occur during the build process, and address them appropriately.

Next, analyze the code for performance bottlenecks by profiling the application to identify slow or resource-intensive areas. Consider factors such as algorithm efficiency, memory usage, and execution time. Implement optimizations where possible to enhance performance.

Then, perform a code smell analysis to detect issues such as duplicated code, overly complex functions, poor naming conventions, and excessive coupling or dependencies. Use static code analysis tools to assist in identifying potential smells, and refactor the code to improve readability, maintainability, and adherence to coding standards.

For each identified issue, document the following:

- Description: What is the specific problem or issue identified?

- Location: In which part of the codebase does the issue occur?

- Impact: How does this issue affect the project's performance, readability, or maintainability?

- Solution: What steps will you take to resolve or mitigate the issue?

Finally, summarize the findings and improvements in a report, providing recommendations for future development practices to prevent similar issues. Ensure that the codebase is thoroughly tested after making changes to confirm that all issues have been resolved and that no new problems have been introduced.

+++ src/tasks/ai_task.cpp
@@ -94,7 +94,7 @@ ai_state_t analyze_behavior(const sensor_data_t* data) {
     
     // Check for high activity states
     if (data->wifi_networks_count >= HIGH_WIFI_ACTIVITY_THRESHOLD) {
-        excitement_level = min(excitement_level + 5, 100);
+        excitement_level = min(excitement_level + 5, (uint32_t)100);
         
         if (excitement_level > 80) {
             return AI_STATE_EXCITED;
@@ -110,7 +110,7 @@ ai_state_t analyze_behavior(const sensor_data_t* data) {
     
     // Learning state when processing data
     if (current_ai_state == AI_STATE_SNIFFING && state_duration > 5000) {
-        learning_progress = min(learning_progress + 10, 100);
+        learning_progress = min(learning_progress + 10, (uint32_t)100);
         return AI_STATE_LEARNING;
     }
     
@@ -140,7 +140,7 @@ void update_learning_metrics(ai_state_t new_state) {
             break;
             
         case AI_STATE_EXCITED:
-            excitement_level = min(excitement_level + 10, 100);
+            excitement_level = min(excitement_level + 10, (uint32_t)100);
             break;
             
         case AI_STATE_SLEEPING:
@@ -155,8 +155,8 @@ void update_learning_metrics(ai_state_t new_state) {
     }
     
     // Cap metrics at reasonable values
-    learning_progress = min(learning_progress, 100);
-    excitement_level = min(excitement_level, 100);
+    learning_progress = min(learning_progress, (uint32_t)100);
+    excitement_level = min(excitement_level, (uint32_t)100);
 }
 
 /**
